src/sleep_disorder/utils/common.py
import os
from pathlib import Path

import yaml
from box import ConfigBox
from box.exceptions import BoxValueError
from ensure import ensure_annotations

from sleep_disorder.logging import logger


@ensure_annotations
def read_yaml(path_to_yaml: Path) -> ConfigBox:
    """reads yaml file and returns ConfigBox.

    Args:
        path_to_yaml (Path): Path to yaml file.

    Raises:
        valueError: if yaml file is empty.
        e: empty file.

    Returns:
        ConfigBox.
    """
    try:
        with open(path_to_yaml) as yaml_file:
            content = yaml.safe_load(yaml_file)
            logger.info(f"yaml file: {path_to_yaml} loaded successfully.")
            return ConfigBox(content)
    except BoxValueError:
        raise ValueError("YAML file is empty.")
    except Exception as e:
        raise e


@ensure_annotations
def create_directories(path_to_directories: list, verbose=True):
    """creating list of directories.

    Args:
        path_to_directories (list): list of directories.
        verbose (bool, optional): ignore if multiple dirs is to be created. Defaults to True.
    """

    for path in path_to_directories:
        os.makedirs(path, exist_ok=True)
        if verbose:
            logger.info(f"Directory created at: {path}")


@ensure_annotations
def get_size(path: Path) -> str:
    """Get's size of the file in KB

    Args:
        path (Path): Path to file.

    Returns:
        str: Size of the file.
    """

    size_in_kb = round(os.path.getsize(path) / 1024)
    return f"~ {size_in_kb} KB."
