requirements.txt
pandas
numpy
matplotlib
seaborn
statsmodels
scikit-learn
notebook
Flask
python-box
PyYAML
ensure
-e .
